Initial code check-in. While there is still a lot of clean-up and niceties to add, the code is very near to feature complete. Multiple card sets can be created, renamed, deleted, or have their "strike-through" data erased. Individual cards are correctly displayed and "strikes" are recorded and restored as the user scrolls through the cards. The user can jump to any card in the set. Manually "striking through" of a passcode copies the value of the passcode to the clipboard so it can be pasted into the form for authentication if needed.
There is a password prompt activity that will sit in front of the main menu if a password is saved (encrypted, of course) in the users preferences. However, there is currently no facility for setting this password, so the prompt is always bypassed for now. Once the settings activity is added, this should become active.

There is no help facility yet; menu items exist but I haven't had a chance to add the help view and help index activities.

As far as I can tell, the database structure and interface code should be mature and should not need to change in the foreseeable future. Therefore, it should be safe to experiment with this app without losing data due to updates. That said, I would recommend keeping a backup of all card set parameters external to the app for now until this is known not to change.