Adjusted the card set logic to limit the number of columns and passcode length based on the available space on the typical Android device screen. If the product of these two values is greater than a certain constant (admittedly found via lots of experimentation), the New Card Activity will not allow the card set to be created.
In the Card View, if the product of the number of columns and the password length exceeds a another constant, the view forces itself to be in landscape orientation. This is because the ScrollView layout object in Android does not do horizontal scrolling, so we can scroll sideways when the card becomes too wide.

Updated the validation code in the Cardset class to deal with the changes above. Also fixed a few bugs in the non-default constructor.

Fixed the CardDBAdapter class to prevent an exception being thrown based on the new limitation specified above.

Adjusted the card set list row layout to make the TextView fill the entire row, so now you can tap anywhere in the row to select that item.