Tweaked the code for determining maximum allowable card width and at what width we force cards to be viewed only in landscape. What I had before wasn't working well, so some more experimentation found better values. My formula also did not take into account the spacing between columns, so that was factored in as well. The logic for determining these two items (max card width and portrait/landscape cut-off) has been moved into the Cardset class. I'm not sure if that's the best place to put it, but it seems to be the most self-contained approach.
Fixed a few minor things during card view generation. There is still a known issue with the forced-landscape mode that causes the "Generating codes" progress dialog to pop up a second time and just sit there in what seems like a stalled state. It's easy to clear my just hitting Back, but that's not necessarily obvious. This needs to be fixed, but I need to find out exactly why it's happening first. I think it has to do with the destroy/restore approach Android does to changing orientation, so I may need to look into Configuration and similar classes.

I also fixed a few bugs in the New Card activity with respect to the new card width rules. Again, it's not perfect; there are issues with the focus change event handlers, so I need to test further to see how to make them behave better.