Added the Card Set Details activity. This read-only view shows all the parameters that make up a given card set, which is handy for displaying this information after the card set has been created. (Currently, there is no provision for editing a card set besides renaming it, deleting it, or scrolling through cards and toggling passcodes.) This activity also includes several of the same actions from the main menu's context menu specific for the selected card set, such as renaming it, deleting it, clearing all toggles, etc.
I think I've fixed the issue with the progress dialog hanging in the Card View Activity when the orientation changes, or when we force landscape orientation for extra-wide cards.  The CardViewState class now holds the current state of a card, including its displayed passcodes and toggle data. When a configuration change occurs (for our purposes, the screen is rotated or we force landscape mode), the state of this data goes into this class, which then persists while the activity itself is destroyed and recreated. Then the persistent data is read and recreating the card get a "shortcut" that skips the progress dialog and the expensive encryption step.