Started adding encryption of sequence keys. If the user sets an access password, PPP will encrypt the sequence keys in the database with a version of AES and store the encrypted key. Each time the card set parameters are written or read from the database, the sequence key will be encrypted or decrypted respectively. The encryption seems to be working fine as far as I can tell, but the decryption isn't. I'm getting bad padding exceptions, so obviously something isn't working as expected. There's still some debugging to do.
To finish the thought, when the user clears the password in the settings activity, this will subsequently decrypt all the sequence keys and store them in the DB in plain text. Obviously if decrypting one sequence key on the fly isn't working, neither is this, so it's not ready for use yet.