Looks like I've fixed the encryption issues surrounding setting the password and encrypting the sequence keys. The problem essentially boiled down to a coding error; encryption occurred OK, but during decryption I was not converting the binary data back to a proper string, and then I had to worry about removing null characters left over from the decryption.  The encryption should now be transparent.  If no password has been set, the sequence keys are stored in the DB in the clear; if a password is set, all sequence keys are encrypted and will be decrypted on the fly when the card set is accessed.
I never got the built-in Java encryption code to work, so I left in the Bouncy Castle code, which is getting added here for the first time. I'm more comfortable with the Bouncy Castle stuff due to my experience with it on Cryptnos, so that's OK.  I updated the About activity to include the BC license info and acknowledgments.

Also added a new, albeit ugly, launcher icon to replace the default one.  I know the shadows are off so they'll need to be reworked, but at least for now it's something different.  A non-Eclipse "icons" directory will include the SVG and Paint Shop Pro sources for these files.