Improved the handling of the encryption/decryption of sequence keys when the master password is set/cleared.  Although this is more obvious on the emulator than on my personal devices, having multiple card sets in the database could result in a lengthy cryptographic process as all the sequence keys are [en|de]crypted in turn.  This isn't a problem when opening the card set details or opening the card view itself; in these cases only one sequence key is being decrypted, so the process is quick enough to be transparent.  This isn't the case for doing the whole database when the password is set or cleared.
To improve the user experience, the actual work of setting or clearing the password and encrypting or decrypting the sequence keys has been moved to a worker thread tied to a progress dialog.  The dialog states that the sequence keys are being encrypted or decrypted and provides a status of how far along the process is.  Tests look good on the emulator, and I'm anxious to see how well this does on an actual device.  My personal devices are probably too fast and I'll use too few card sets to really notice the improvement, but it could offer a much better experience for lower end devices.

I also started the process of readying PPP for Honeycomb by bumping up the target API and taking care of a few known "gotchas" I picked up from Cryptnos (like the deprecation of the ClipboardManager class).  PPP should now use the default Honeycomb "holographic" theme, but there are no dedicated layouts yet.

I improved the launcher icons and saved copies of the Paint Shop Pro intermediate files.

I updated the project URL in all of the header comments to point to the Google Code site, rather than the GPF site (which is unrelated to this project).