More layout tweaking. After the previous changeset, I discovered that the new dimension properties no longer worked well on my personal Motorola Droid. This was because the passcode text size dimension, specified in scaled pixels, ended up with much larger text on the Droid screen than it did on the baseline HVGA Android 1.6 screen in the emulator. After a bit of experimentation, I added a "hdpi" resource for the dimensions, which the Droid and similar WVGA devices will use.
In the process, I decided to see how well PPP fares on an "ldpi" or "small" screen. It... did not go well. Using the default resource (same as the baseline above), the passcodes were unreadable and likely would be nearly impossible to toggle. Unfortunately, there does not seem to be an easy solution to this problem. I cannot increase the text size without requiring the ability to scroll, which is currently not possible. This means we will be unable to practically support "small" screen sizes. I have added the appropriate filter to the manifest and I'll update the Google Code wiki to reflect this change.