Because I can't leave well enough alone, I've already got a working solution to the small/ldpi screen situation. I have added a new SinglePasscodeActivity class that provides a simplified interface to display a single passcode at a time. Like many other PPP clients, this lets the user specify the card, column, and row identifiers and generates the passcode on the fly. The passcode is still a toggle button and its state is still recorded in the database, just like the main card view. This view attempts to jump to the passcode immediately after the last toggled passcode on the specified card or, if the last passcode was toggled, it jumps to the first passcode on the next card. The user can, of course, override this by selecting the values directly from the list.
I added a res/values-small/dimensions.xml resource and in it created a small boolean value called "is_small_screen". This is set to true in values-small and false in all the other dimension files. The main menu now checks for this, and if a card set is selected and the device has a small screen, it gets the simplified view. If the device does not have a small view, the default card view is displayed.

There's probably still a lot of clean-up and debugging to do, but it seems to be working so far. Looks like I'll have to remove the changes to the wiki, or at least update them with some better explanatory text.